requirements.txt
azure-identity>=1.16.0
openai>=1.37.0
python-dotenv>=1.0.1
requests>=2.31.0
